requirements/base.txt
# Base requirements for all environments

# Django
# ------------------------------------------------------------------------------
Django==4.1.9
django-environ==0.10.0
django-model-utils==4.3.1
django-allauth==0.54.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-redis==5.2.0
django-cors-headers==4.0.0

# Django REST Framework
# ------------------------------------------------------------------------------
djangorestframework==3.14.0
dj-rest-auth==4.0.1
drf-spectacular==0.26.2

# Database
# ------------------------------------------------------------------------------
psycopg2-binary==2.9.6

# Media and Files
# ------------------------------------------------------------------------------
Pillow==9.5.0

# Task Queue
# ------------------------------------------------------------------------------
celery==5.2.7
django-celery-beat==2.5.0
flower==1.2.0

# ASGI Server & WebSocket
# ------------------------------------------------------------------------------
daphne==4.0.0
channels==4.0.0

# Utilities
# ------------------------------------------------------------------------------
python-slugify==8.0.1
pytz==2023.3
Babel==2.12.1

# Invoice Management Dependencies
# ------------------------------------------------------------------------------
fpdf2==2.7.4
requests==2.31.0
pandas==2.0.3
tabulate==0.9.0

# Utilities and Misc
# ------------------------------------------------------------------------------
redis==4.5.5
hiredis==2.2.3
django-admin-sortable2==2.1.5
django-dynamic-preferences==1.15.0
